requirements.txt
Flask
langchain-community
langchain-qdrant
qdrant-client
nltk
python-dotenv
